--- tests/Digdir.Domain.Dialogporten.WebAPI.E2E.Tests/Features/V1/ServiceOwner/Dialogs/Commands/Create/Snapshots/CreateDialogSnapshotTests.Create_Dialog_DialogId_Conflict_Verify_Snapshot.verified.json
@@ -1,13 +1,0 @@
-﻿{
-  "type": "https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.8",
-  "title": "Conflict.",
-  "status": 409,
-  "instance": "/api/v1/serviceowner/dialogs",
-  "conflictingDialogId": "00000000-0000-0000-0000-000000000000",
-  "errors": {
-    "IdempotentKey": [
-      "\u002700000000-0000-0000-0000-000000000000\u0027 already exists with DialogId \u002700000000-0000-0000-0000-000000000000\u0027"
-    ]
-  },
-  "traceId": "00-00000000000000000000000000000000-0000000000000000-00"
-}
