+++ tests/Digdir.Domain.Dialogporten.WebAPI.E2E.Tests/Features/V1/ServiceOwner/ErrorResponse/Snapshots/ErrorResponseSnapshotTests.Should_Return_409_For_Duplicate_Dialog_Id.verified.json
@@ -8,5 +8,6 @@
     "IdempotentKey": [
       "\u0027duplicate-idempotent-key\u0027 already exists with DialogId \u002700000000-0000-0000-0000-000000000000\u0027"
     ]
-  }
+  },
+  "conflictingDialogId": "00000000-0000-0000-0000-000000000000"
 }
