+++ tests/Digdir.Domain.Dialogporten.WebAPI.E2E.Tests/Features/V1/ServiceOwner/ErrorResponse/Snapshots/ErrorResponseSnapshotTests.Should_Return_409_For_Update_With_Duplicate_Transmission_Idempotent_Key.verified.json
@@ -8,5 +8,8 @@
     "IdempotentKey": [
       "Duplicate IdempotentKey detected in dialog transmissions. Conflicting keys: \u0027duplicate-transmission-key\u0027."
     ]
-  }
+  },
+  "conflictingIdempotentKeys": [
+    "duplicate-transmission-key"
+  ]
 }
