+++ tests/Digdir.Domain.Dialogporten.WebAPI.E2E.Tests/Features/V1/ServiceOwner/ErrorResponse/Snapshots/ErrorResponseSnapshotTests.Should_Return_409_For_create_With_Duplicate_Transmission_Idempotent_Key.verified.json
@@ -3,13 +3,13 @@
   "title": "Conflict.",
   "status": 409,
   "instance": "/api/v1/serviceowner/dialogs/00000000-0000-0000-0000-000000000000/transmissions",
-  "conflictingIdempotentKeys": [
-    "00000000-0000-0000-0000-000000000000"
-  ],
+  "traceId": "00-00000000000000000000000000000000-0000000000000000-00",
   "errors": {
     "IdempotentKey": [
       "Duplicate IdempotentKey detected in dialog transmissions. Conflicting keys: \u002700000000-0000-0000-0000-000000000000\u0027."
     ]
   },
-  "traceId": "00-00000000000000000000000000000000-0000000000000000-00"
+  "conflictingIdempotentKeys": [
+    "00000000-0000-0000-0000-000000000000"
+  ]
 }
